fix(api): fix connection to mongodb
